[KYV-649] Added support to heck if user has come from our own customer page or redirected straight to paymentMethods page
