+++ src/enums/RouteConfigs.ts
@@ -1,0 +1,3 @@
+export enum RouteConfigs {
+    FROM_CUSTOMER_DETAILS_ROUTE = "fromCustomerDetailsRoute",
+}
