+++ src/i18n/sv.json
@@ -2,6 +2,7 @@
   "common": {
     "page-title": "Helsingfors nätbutik kassa",
     "cancel-and-return": "Föregående",
+    "cancel-and-return-referrer": "Sluta beställa",
     "language": {
       "language-selector": "Språkväljare",
       "fi": {
